# public/swagger.yaml
openapi: 3.0.3
info:
  title: Dentista Now API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: Dentista Now
    url: https://dentista-now.vercel.app
servers:
  - url: https://dentista-now.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
...
paths:
...
  /agendamentos/criar:
    post:
      tags:
        - Agendamentos
      summary: Criar novo agendamento
      description: |
        Cria um novo agendamento com sistema de rodízio automático.

        **Sistema de Rodízio:**
        - Se barbeiro_preferido informado: Agenda com ele (se disponível)
        - Se não informado: Usa rodízio balanceado (menos agendamentos do dia)

        **Validações:**
        - Verifica conflito de horário
        - Sugere próximos horários se ocupado
        - Dispara webhooks global e personalizado do barbeiro

        **Exemplo de uso:**
        ```bash
        curl -X POST 'https://dentista-now.vercel.app/api/agendamentos/criar' \
          -H 'Authorization: Bearer SEU_TOKEN' \
          -H 'Content-Type: application/json' \
          -d '{
            "cliente_nome": "João Silva",
            "telefone": "11999999999",
            "data": "2025-12-20",
            "hora": "14:30",
            "servico_ids": ["uuid-servico-1"],
            "barbeiro_preferido": "Hiago"
          }'
        ```
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cliente_nome
                - telefone
                - data
                - hora
                - servico_ids
              properties:
                cliente_nome:
                  type: string
                  description: Nome do cliente
                telefone:
                  type: string
                  description: Telefone do cliente
                data:
                  type: string
                  description: Data do agendamento (YYYY-MM-DD ou DD-MM-YYYY)
                  example: "2025-12-20"
                hora:
                  type: string
                  description: Hora de início (HH:MM)
                  example: "14:30"
                servico_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Array de IDs dos serviços
                barbeiro_preferido:
                  type: string
                  description: Nome ou UUID do barbeiro preferido (opcional)
                observacoes:
                  type: string
                cliente_id:
                  type: string
                  format: uuid
                  description: ID do cliente existente (opcional)
            examples:
              com_barbeiro:
                summary: Agendamento com barbeiro específico
                value:
                  cliente_nome: "João Silva"
                  telefone: "11999999999"
                  data: "2025-12-20"
                  hora: "14:30"
                  servico_ids: ["uuid1", "uuid2"]
                  barbeiro_preferido: "Hiago"
              rodizio:
                summary: Agendamento com rodízio automático
                value:
                  cliente_nome: "Maria Santos"
                  telefone: "11988888888"
                  data: "2025-12-20"
                  hora: "15:00"
                  servico_ids: ["uuid1"]
      responses:
        '201':
          description: Agendamento criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      agendamento_id:
                        type: string
                        format: uuid
                      barbeiro_atribuido:
                        type: string
                      data:
                        type: string
                      horario:
                        type: string
                      valor_total:
                        type: number
                      duracao_total:
                        type: integer
        '409':
          description: Conflito de horário
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      barbeiro:
                        type: string
                      horario_solicitado:
                        type: string
                      sugestoes:
                        type: array
                        items:
                          type: string
...